Allow specifying fsc path via environment variable
If the F# installation path can't be found via the usual registry keys, the
environment variable FSHARPINSTALLPATH will also be tried before giving up.
Also removed code paths for dealing with querying the registry on very old
versions of Windows.

Fixed some whitespace issues and added a hint in .gitattributes so that GitHub
doesn't think this is a C# project.