Install assembly to Program Files as well as GAC
We have to specify the component twice: once to get a copy registered in
the GAC, and again so that we can install a copy of the assembly to
Program Files.  The former is for the .NET runtime's use and is required
because we also register it as a CodeDom provider, and the latter is for
the benefit of developers so that the assembly shows up in the VS "add
reference" window.