Fix SimpleInstaller on Mono; dedupe installer code
Mono's installutil refused to work with an Install class written in F#,
so it has been replaced with a C# version, which for whatever reason,
Mono's installutil has no trouble running.

The machine.config installer code was common to both the MSI and simple
installers, so it has been broken out into a library.